+++ src/main.js
@@ -2,7 +2,7 @@ import Vue from 'vue'
 import vuetify from './plugins/vuetify'
 import App from './App.vue'
 import router from './router'
-import i18n from './i18n'
+import i18n from './plugins/i18n'
 
 Vue.config.productionTip = false
