+++ src/plugins/i18n.js
@@ -4,7 +4,7 @@ import VueI18n from 'vue-i18n'
 Vue.use(VueI18n)
 
 function loadLocaleMessages () {
-    const locales = require.context('./locales', true, /[A-Za-z0-9-_,\s]+\.json$/i)
+    const locales = require.context('/src/locales', true, /[A-Za-z0-9-_,\s]+\.json$/i)
     const messages = {}
     locales.keys().forEach(key => {
         const matched = key.match(/([A-Za-z0-9-_]+)\./i)
